Fix incorrect store action when UnsafePass writes to a MSAA target
Fix incorrect behavior identified in https://jira.unity3d.com/browse/UUM-67135

The NRP Compiler logic detects if the current render target is being used outside the current native render pass (e.g. when the pass is broken up by an unsafe pass), and determines the store action for this case. For MSAA resources, the logic loops through the Readers (which contains the list of passes that either read or write to the resource), and it assumed that MSAA Samples need to be preserved only if it is bound as a fragment (Raster Pass render target), or if the `bindMS` flag on the texture is set. This logic did not cover the use case of 2D Renderer, where the MSAA texture was used inside an Unsafe Pass by as argument to SetRenderTarget. 

This PR does ~two~ one thing:
1) ~Removes the workaround 2D had to use to get around this issue (cc @kenny-tan)~
2) Improves the underlying issue by always using StoreAndResolve action when resource is read by an Unsafe Pass. We must currently store both surfaces because we cannot know if the Unsafe Pass needs the resolved or non-resolved data. (cc @charles-hollemeersch, @adrien-moulin)
    - **PLEASE NOTE!** This may potentially result in worse performance

After the fix, RG Viewer now shows "StoreAndResolve" instead of "Resolve": 
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/b36189b0-5da4-4405-aab9-be4e6d4116e0)

**NOTE** ~There is a separate bug where the "store reason" is incorrect - new RG Viewer doesn't correctly handle the separate store reasoning for MSAA data, I will handle this in a separate PR.~ Fixed in https://github.cds.internal.unity3d.com/unity/unity/pull/46792