Enabled Foveated Rendering for shaders on Vision OS and implemented for soft particles
*This PR partly addresses: https://jira.unity3d.com/browse/XRVOSB-20

In the original issue, particles that have the Soft Partilces feature enabled would appear wildly distorted on the Vision Pro. Soft particles perform a sample of the depth buffer in order to neatly fade their texture when close to other objects. Previously, when foveated rendering was enabled, the shader would sample the depth buffer at the wrong coordinates, resulting in a heavily distorted image. 

This PR allows shaders to use Foveated Rendering code when on the Vision OS platform, and modifes the particle shader to properly take foveation into account when querying the depth buffer.