[HDRP] Hair - Fix HQ line minor regression
This PR fixes 3 minor regression bugs related to HQ lines (2 of which are https://jira.unity3d.com/browse/UUM-58687, https://jira.unity3d.com/browse/UUM-58614). 

The third bug is checking just one end of a segment for visibility is too crude and will make hair around silhouettes disappear. The change done here is to check both ends, which fixes the problem for all cases we've observed (but is not a complete solution). A more complete solution for this would be to do a test against depth pyramid. This is a regression and the logic used to check both ends of the segment