Docs quality sprint 17 April, Oleks
Docs quality sprint 17 April, Oleks

Tickets in this PR:

DOCG-5296. "a combination of 2D and 3D Lights and 2D and 3D Renderers in a single Scene"

DOCG-4772. Unlit shaders edit.

DOCG-5083 cameraOutput is not a property of UniversalAdditionalCamera… 

DOCG-4935 There should be {} at the end of the property after "white"

DOCG-5449. Spelling error.