[VFX] Fix exposed properties reset when editing multiple VFX at the same time on inspector
When selecting multiple VFX with exposed properties, if those exposed properties had different values, any change in the inspector was resetting the values of the exposed properties, because we were not tracking GUI changes locally.

This issue was also responsible of the Allow Instancing property not being properly applied when selecting multiple objects.

This PR also adds a property drawer for animation curves, that were not displayed if multiple objects with different exposed animation curves were selected.
<img width="533" alt="image" src="https://media.github.cds.internal.unity3d.com/user/2805/files/f6c2c4e1-a4b5-4cc4-8793-c2acd541ae11">