[2022.3 Backport] [URP] [Native Render Pass]Fix incorrect RenderTextureDescriptor when rendering into a texture and having a renderScale
When Render scale != 1 and we are rendering to a Render Texture. The NativeRender pass was creating the RenderTextureDescriptor with an invalid dimensions.

Make sure that Attachements sizes are the same as pass sizes.

This is a backport from @alex-vazquez fix:
[Original PR](https://github.cds.internal.unity3d.com/unity/unity/pull/48301)
[Original Jira Bug](https://jira.unity3d.com/browse/UUM-61468)

This PR fixes:
https://jira.unity3d.com/browse/UUM-58611
https://jira.unity3d.com/browse/UUM-59081

Port Jira tickets: [UUM-69058](https://jira.unity3d.com/browse/UUM-69058) and [UUM-69115](https://jira.unity3d.com/browse/UUM-69115)