Rewrite Alpha To Coverage Alpha Clipping Logic
This change rewrites a good deal of the logic within URP's alpha clipping function that's responsible for handling alpha to coverage behavior. In addition to handling the degenerate use case where cutoff is set to zero, this code now also handles the case where alpha and cutoff are equal. It does this by modifying the sharpening function to yield 1.0 rather than 0.5 in the equal case while still maintaining the other parts of the curve.