Fix issues when inspecting renderer and volume profile assets under specific circumstances
This PR fixes two issues:

https://jira.unity3d.com/browse/UUM-71897 
Regression from recent PR https://github.cds.internal.unity3d.com/unity/unity/pull/48923 caused errors to be printed if URP is not the active pipeline when inspecting URP renderer assets. Added a null check.

https://jira.unity3d.com/browse/UUM-71890
When opening the editor, it's possible SRP is configured to be used but it has not been activated. In this situation you may see an empty volume inspector. Adding a message to let user know what is happening.