[Port] [2021.3] UUM-70600: Cookie looks corrupt in editor when Android is current target platform
This fix works around a shader miscompilation when targeting mobile. In practice this causes artifacts in code that uses octahedral encoding/decoding functions.

A bit of history is warranted here. In this PR https://github.cds.internal.unity3d.com/unity/unity/pull/46787 I fixed a similar issue, and the explanation in that PR still holds. That fix addressed a case where we call the octahedral encoding function, which expects a float as input, with a half on mobile. I 'fixed' it by making the function take half's on mobile. Unfortunately, that fix introduced a regression in cookie shading. The cookie issue is sort of the opposite - we are passing floats where the function expects half on mobile.

To make it completely clear, both of these setups _should_ work - half<->float conversion should be happening without issues. I'm convinced we are dealing with a miscompilation here. In order to make a fix that works for both kinds of inputs, both half and float, I drilled into the shadercode to find the minimal repro. The problematic instructions are a few calls to abs() which give incorrect outputs. When the abs() is present, instead of half<->float conversion, we seem to get something more akin to a direct bit-cast between the 2 datatypes, which results in reading uninitialized memory when converting half->float (hence why the repro in https://github.cds.internal.unity3d.com/unity/unity/pull/46787 is violently shaking), and truncation of bits when converting float->half (hence why the repro in the cookie bug is completely static, but corrupt looking).

**As a fix**, I've reverted the function back to its initial state of taking floats as input, and replaced the abs() calls with a manual sign check. This fixes both of the aforementioned issues.