Graphics/SRP/RPF - [UUM-57147] - Disabling NRP for Editor FinalCopyDepth pass
This PR disables Native Render Pass (NRP) support of URP Editor-only `FinalCopyDepth` pass which is used a pre-step for the Post-Process Gizmos pass in URP 3D. Using NRP in this pass triggered a corner case which is not well covered by NRP pipeline in Compatibility mode. Indeed, NRP pipeline begins a native render pass with a number of MSAA samples taken from the `CameraTargetDescriptor`. `CameraTargetDescriptor` seems to obtain that information from the RenderPipeline asset, so depending on what users select in the UI, it can be no MSAA/x2/x4/x8. On the other hand, `FinalCopyDepth` seems to be the only pass targeting `k_CameraTarget` in URP 3D and its attachment (backbuffer) has no MSAA enabled, triggering a mismatch in NRP validation layer as the number of samples is different (unacceptable in NRP). As this is Editor only + Compatibility mode, I assume that using NRP for this pass is not really important and can be disabled. Compatibility mode with no NRP works fine for this case.

Note that NRP Render Graph should be preferred over NRP Compatibility mode for Unity 6.