[APV] Fix leak reduction mode in complex cases
Updated the leak reduction mode to Performance/Quality options
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/e643a61d-a90a-48a4-980e-0fb6140d285c)

Performance is the same as previous ValidityBased mode. Does a single sample with warped UVW
Quality is a new one that does up to 3 samples and never samples an invalid probe to eliminate leaking.

See improvement on the difficult case of the tent in the URP sample scene
| Performance | Quality |
| -- | -- |
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/12b7105f-00f3-4f31-b82d-f4d928725fae) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/2fa02f0a-2e2b-455a-83b9-b919b7188f3a) |