Reduce Banding on FSR Upscaled Render Target
The previous code was using the same graphics format for the scaling
setup and FSR upscaled RTs. When FSR is active, we use a unorm format
on the EASU input RT (scaling setup's output) to avoid banding since
its content is gamma2 encoded.

We were unnecessarily using this format for the final upscaled RT as
well which contains linearly encoded color data and can safely use
the pipeline's standard color format.

In this case, the unnecessary format modification was making the
banding more visible.

Before:
![image](https://media.github.cds.internal.unity3d.com/user/4435/files/6107cf5e-e26e-4590-96ae-15e151bee4dd)

After:
![image](https://media.github.cds.internal.unity3d.com/user/4435/files/7e595fa3-bd24-4bf2-b124-ef0966e44ed5)