Update pipelines with graphicsJobs setting enabled to use command line argument
It was brought up that we can enable the `graphicsJobs` setting using the `-force-gfx-jobs split` command line argument, therefore removing the need of using the `com.unity.cli-project-setup` package.

This PR adds the argument to the utr flags and removes the call to open the project and execute the method from the `com.unity.cli-project-setup` package.

The pipelines run with `kGfxThreadingModeSplitJobs` instead of `kGfxThreadingModeNonThreaded`