UUM-65454 MaxSubpass discrepancy with RenderPassSetup and NativePassCompiler (amelioration)
JIRA: https://jira.unity3d.com/browse/UUM-65454

When merging subpasses together in a NativePass, we don't check if there will be more than 8 native render subpasses. This PR adds a check for this when verifying if a certain graph pass can be merged into a Native Pass, as well as a test. If merging the graph pass would result in too many native subpasses, we simply create a new Native Pass

This is the second PR for this fix, the first one was more of a hotfix because of the time constraint, this is a more involved refactor that should speed up the first fix

first PR: https://github.cds.internal.unity3d.com/unity/unity/pull/46620