Shader Graph Sample - Production Ready Shaders
Adds a new set of samples - Production Ready Shaders - to the Shader Graph package.

The Shader Graph Production Ready Shaders sample is a collection of Shader Graph shader assets that are ready to be used out of the box or modified to suit your needs.  You can take them apart and learn from them, or just drop them directly into your project and use them as they are. The sample also includes a step-by-step tutorial for how to combine several of the shaders to create a forest stream environment.

![thumbnail](https://media.github.cds.internal.unity3d.com/user/4330/files/e77ac010-f157-450e-972e-17c8b1d1e45d)

Documentation:
https://docs.google.com/document/d/1vAwWO39TUg8VTlzznlKHBmwYftgHfFZ7eoLJNd_DhIU

Blog Post (draft):
https://docs.google.com/document/d/1F8yArILrt19uNWzd0yfu59KmjuRjUOxQo9Y84-y6Zao

Video to announce the content:
https://drive.google.com/drive/folders/1T2Y4x3gGAJmA1ikQv5485CTRganBRLfg
https://docs.google.com/document/d/1WxKL3xZ0GV84lgV3VoCcP_5n6mxpo87gPeoUpI9F7tA

Jiras associated with this project:
https://jira.unity3d.com/browse/GSG-822
https://jira.unity3d.com/browse/PLATGRAPH-2851
https://jira.unity3d.com/browse/PLATGRAPH-3045
https://jira.unity3d.com/browse/PLATGRAPH-3046
https://jira.unity3d.com/browse/PLATGRAPH-2793
https://jira.unity3d.com/browse/PLATGRAPH-2205