Graphics/SRP/RPF - [UUM 69082][URP][RG][Camera Stacking] Game view background turn yellow after enable render graph
When more than two cameras are stacked in 2D with URP and RG activate, the `clearOnFirstUse` is set to `false` for all the overlay cameras, and `true` for the Base camera.
This flag is allocated in the `TextureDesc.clearBuffer` of the `TextureResource`.
When the hash of the Render Graph is created, this flag was not added to the hash of the graph.
So, the RG was returning the same cache hash for base camera, and overlay cameras, and it was clearing the color buffer as it was indicated in the base camera for all the overlay cameras.
I have added the `clearBuffer` to the compute of the hash, and to be sure for the future also the `discardBuffer` as this is another flag in the camera stacking.