Enable RuntimeProfilerGivesNonZeroOutput, Remove Hardcoded string from SRP Recipes
[GFXFOUND-327](https://jira.unity3d.com/browse/GFXFOUND-327)
Enables a disabled test on certain platforms -- looking at the previously ignored platforms:

- [Passes on Linux](https://unity-ci.cds.internal.unity3d.com/job/37741645/logs/execution?line=3032&q=Profiler)
- [Passes on Windows DX12](https://unity-ci.cds.internal.unity3d.com/job/37741655/logs/execution?r=37741655&line=6948&q=Profiler)
- [Passes on Windows DX12 XR](https://unity-ci.cds.internal.unity3d.com/job/37741650/logs/execution?r=37741650&line=3147&q=Profiler)
- [Ignored on WebGL](https://unity-ci.cds.internal.unity3d.com/job/37741651/logs/execution?line=3807&q=Profiler)
- [Ignored on Quest](https://unity-ci.cds.internal.unity3d.com/job/37741657/logs/execution?r=37741657&line=3330&q=Profiler)

This test was disabled by hardcoding a string into the TestFilter argument of recipes which overrode the test filter given through Yamato UI etc. I would consider this an unfavorable practice as it's a quite surprising way of disabling a test. It is also blocking a certain effort to bring test filtering to the Graphics Test Framework.

Unfortunately, I could not find any way of disabling it on Quest **only** so I had to disable it on all Android platforms.

Therefore, effectively this PR enables the test on Windows DX12 and Linux but disables it on Android devices including Quest.