NativeRenderPass in URP can reach a hard-coded limit that breaks the rendering when using Compatibility Mode
When extending URP with custom renderer features, and Native RenderPass is enabled in the non-RenderGraph pass, the renderer can crash because of a hard-coded value `kRenderPassMaxCount = 20` used by an static array to set his size.

That static array is saving in the index of the render Pass the hash of the render pass. That is causing that when the index of the pass is higher or equal than 20, the array has no place to save the hash of the pass. In fact, most of the places of the array are empty, because the hash is only allocated when the pass is active.
With this change, the static array is changed to a Dictionary with the key as the index of the Pass, and the value as the hash of the Pass, so when the index of the RenderPass is higher than 20 this will no cause an `IndexOutOfBounds` error.

The rest of structures used in the class could still be working without any problem

Jira ticket: [UUM-69943](https://jira.unity3d.com/browse/UUM-69943)