[HDRP] Adding SG custom refraction example to Transparency sample
This concludes the task of adding transparent samples to HDRP. The last needed part was the shader graph example which needed to include compute thickness and new nodes added by @soufiane-khiat a while ago. 

This adds 3 examples: 
- The gummy bears, which shows two ways of using compute thickness by rendering gummy bears and its x-ray using the same shader. 
- The lens effect which has been snatched from the dead Werner demo
- And a full custom refraction shader which replicates the sphere refraction model in HDRP using an Unlit shader. 

Those should be good base for user to kickstart a custom refraction effect in HDRP

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/0565c42d-b16d-4504-91a0-8056d440e8c4)

this adds two new thing to check in the HDRP asset when importing a sample. 

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/09c823b1-39b6-46bc-94fc-226e071fcefc)

Lastly, took the opportunity to reduce filesize by replacing the tif with png file which minimize the additional file needed for this addition. 

before: 
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/610bd172-ce4d-408a-a157-347ed9089b40)

now:
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/88c9e5e7-7241-472a-aabc-0f68b23730f4)


Regarding the raytracing quality node. It's not really possible to have a migration / upgrade. But by default, if the user doesn't re-save their graph the result stay unchanged. Then, if they want to upgrade, they need to delete / re-add the keywords to be able to get the new behavior. Since, it's most probably not used a lot (we never had a report regarding this), it's probably acceptable to have this behavior.