UUM-69940 URP NRP RG uses more GPU bandwidth than URP non-RG NRP
JIRA: [UUM-69940](https://jira.unity3d.com/browse/UUM-69940)

## Changes:
It was determined that URP NRP RG uses more GPU bandwidth than URP non-RG NRP (see JIRA task for documentation + results), which is a bug.

This PR removes the gBuffer as a default global, which should reduce the bandwidth usage and is a step towards preventing the resource and texture properties to be out of sync, which is another bug. Considering that this is a breaking change for some users, we also add a sample to show how to re-enable the gBuffer if needed.

The main changes are that we use builder.UseTexture(gBuffer[i]) rather than builder.UseAllGlobalTextures(true) to reference the gBuffer in a pass. This means that we need to call material.SetTexture() in passes that use the gBuffer because it cannot be accessed from the globals anymore.

Some tests were modified as well to use the non-global gBuffer

## Sample:
The sample is a ScriptableRendererFeature that can be added to the rendering asset that will display the contents of the gBuffer and set the gBuffer components as globals after the pass using the builder.SetGlobalTextureAfterPass method. It aims to show (1) how to use the gBuffer as a non-global in a pass and (2) how to set the gBuffer as global (and which names are used internally by URP)