[PLAT-8404][UUM-72560][UUM-72565] Change default URP depth format to D24_S8 on embedded platforms
Changes the default depth format on the embedded platforms to D24_S8. D32_S8 causes issues on some devices and setting it as a default to D24_S8 is a more robust and sane default for our platforms.
This also fixes rendering only a black screen on the NXP i.mx8 device on EmbeddedLinux and QNX with a URP test project.