[HDRP] Adding Environment samples
This sample puts multiple environment effects together in the same scene with different lighting conditions.

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/81df0ebe-0d07-4e4a-851f-83797b2c12b4)


The scene includes examples focusing on the following effects:
- Volumetric Clouds: Demonstrates simple, advanced and manual mode. 
- Lens Flare: How to update SRP Lens Flares data at runtime.
- Night Sky: How to add stars and northern lights.
- Time Of Day: Full 24h cycle with sun, moon and proper shadow map swapping.

More information on this on the Sample showcase inside the sample. 

This PR changes slightly the behavior of atmospheric scattering, it now takes shadow map into account.

When making the PR, we realized it would be nice to take into account to wrap mode of the cloud texture (to clamp instead of automatic repeat some advanced or manual cloud map for example). We'll see if we have time to do it in this PR. 

Docs has also been updated to exaplain visually what is a cloud LUT and cloud map in manual mode. 

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/8d4af8fb-e07c-49a6-b3be-4ad0a9d5cd7c)
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/d9bbbeae-d625-4e38-8576-5760e1074254)

In an ideal world it would have been nice to include a snippet to update cascade of dir light at different intervals but a bug was found ([UUM-72086](https://jira.unity3d.com/browse/UUM-72086)) when including it and we need to wait until it's fixed before including it.