[HDRP] Refactor clouds and water out of HDRenderPipeline
Moved water and clouds as separate systems instead of being part of HDRenderPipeline class. This helps modularizing the code and simplifies dependencies.
Additionally, moved each resources to per system classes that can be stripped independently of the rest of the pipeline resources if the feature is not used at build time. As a result, they show in their own section in the global settings:
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/85471fac-567a-427d-8e56-66feb54b7b57)