Fix URP render graph pass names
Fix bugs with URP RenderGraph pass names. Specifically the DrawObjectPass does not pass the more descriptive pass name to differentiate between Opaques and Transparent. While investigating this, I noticed that the profiler name and the pass name differ in a range of passes. This PR also consolidates the passnames between the profiler, frame debugger and render graph viewer. 

Also, the passes are named very inconsistently. In this PR, I've updated the names to be consistent:
- No use of "Pass" as the end of the pass name, it does not add anything. 
- Are an action, so "Do Something", not "SSAO"
- Consistent over RG viewer, framedebugger and profiler
- Descriptive while short
- Can be camelcase or with spaces, the viewers will "nicify" them (to limit the scope and changes of this PR)

Consistent action name
- "Draw" draws scene objects/meshes, or generally when calling cmd.DrawRendererList in the pass
- "Blit" performs a blit to a resource (so Blit Color LUT instead of Color LUT, Blit SSAO instead of SSAO, ..)
- "Set" sets keywords or global properties

Before
![image](https://media.github.cds.internal.unity3d.com/user/1031/files/c77694bf-9adf-408d-ade3-40915169925e)

After
![image](https://media.github.cds.internal.unity3d.com/user/1031/files/f9c92437-5e98-426b-8745-45daf1b1b7c5)

Consistent over RG viewer, framedebugger and profiler:
- The ProfileSampler names can be camelcase or with spaces.
- The pass names are identical to the profilerSampler names
- The Profiler "nicifies" the names already, so in this PR we adopt that same practice in the RG viewer. The viewer/tool is responsible for this nicification. 

The user can now easily identify a certain pass in both tools.

![image](https://media.github.cds.internal.unity3d.com/user/1031/files/5bd08c41-3191-4db3-8450-f330dc863766)

Also fixed a number of incorrect ProfilerSample scopes. These should not be used in the static ExecutePass function that is shared between RG and non-RG, since in RG the profiler scope is already set automatically. That also removes the need for static ProfileSamplers.