[Port] [2021.3] Fix meta for terrains in HDRP
https://jira.unity3d.com/browse/UUM-71171

The meta pass of the default HDRP terrain shader (TerrainLit.shader) is currently broken. It always outputs black albedo, resulting in no bounced lighting whatsoever. This was a regression introduced in this PR https://github.cds.internal.unity3d.com/unity/unity/pull/20840, which moved around some shader includes, but neglected to update the Terrain shader according. The bug is essentially that the meta pass is incorrectly using camera-relative rendering. This fix is to restore the missing include in the terrain shader.