Fix HDRP Samples issues
Recently, there's been some changes leading to a bug when the sample showcase tab content ended up superposed on the inspector content. 
The simple fix is to remove the sampleshowcase window which has very little purpose and will lead to less potential issue in the future. Now we just select the sample showcase object to make sure the text is displayed at first on scene opened. 

This is what happened when switching to inspector tab when sample showcase was opened when swthcing scenes. 
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/aeae8bd3-9075-44ba-b308-f5c0fd479285)

Also fix atmospheric scattering not sampling shadow map with correct location.
Fix sampling of sun light absorption in volumetric clouds which was done two times and resulted in incorrect lighting in some cases due to lerp