Fix Tonemap artifacts in Scene View
JIRA ticket: https://jira.unity3d.com/browse/UUM-55852

It mentions `Shader Graph + Scene View` only, but it's definitely not related. The issue is because of the Tonemapping system and the `Neutral` mode, when used with 64-bit textures. It's easy to reproduce the issue in the `Game View` too by selecting 64-bit precision in the `Render Pipeline Asset` (`Quality` section). 

This PR fixes the artifacts when using `Neutral Tonemapping` with 64-bit textures. 
Clamping negative values to zero fixes the issue since the `Neutral Tonemapping` curve is not designed to handle negative input values (in comparison to the ACES mode, which handles negative values).

It also updates the `091_ToneMappingNeutralLDR` test we have. The previous `Expected` image seems wrong, some color are off. With the fix, they are fully rendered, respecting the initial image "synthetic chart" `exr` (c.f screenshot below).

![image](https://media.github.cds.internal.unity3d.com/user/1911/files/bd9fd253-6b77-4766-98c7-07d5ad5ce783)

![image](https://media.github.cds.internal.unity3d.com/user/1911/files/4258a254-e588-484e-a891-e536ebae0118)