[VFX][Tests] Added test visualizing the baked SDF
This PR adds a test for the SDF Baker to explicitly visualize the result of the baking, instead of relying only on collisions with SDF like in the test scene 36_SkinnedSDF.

Adding the test highlighted a failure on Apple Silicon, due to restrictions on read-write textures. The PR modifies the baker to avoid problematic read-write operations by adding another copy of the "ray map" and using a ping-pong approach.