Fixed "System.IndexOutOfRangeException: Index -1" in HDRP when zooming out in the scene view
Fixed an out of range exception in HDRP when reaching the max amount of shadows on screen.
This happens when reaching the maximum number of shadows on screen because the request for room in the atlas would then fail and return -1 which was not handled well later down the line.