[21.3] [HDRP] Fix decal material validation
https://forum.unity.com/threads/any-idea-why-decals-turn-to-red-in-2021-3.1371561/

Repro seems to be something like:

create a project with an old version of hdrp and create a decal material with for example affect emissive disabled
close the project, and reopen it with the latest HDRP version
you should have the popup telling you materials need to be upgraded, click yes and wait
after the upgrade, the decal material should have emissive enabled until you open the material in the inspector
This is due to HDRP not being fully initialized yet when the material auto upgrade kicks in, so the correct subshader is not used when checking for enabled passes
Fix consists on not doing anything if we detect we are in this case. This may cause an issue later if we have an upgrade step that relies on reading the passes but since it's not the case currently it should work fine for now.