[HDRP] Fix lens flare screen space and bloom frame settings confusion.
In [this forum post](https://forum.unity.com/threads/disabling-post-processing-effects-in-the-default-volume-profile.1545362/#post-9891525), a user complained that even if Bloom frame settings was set to false, if there was a screen space lens flare override, we "reactivated" the bloom pass without the user's knowledge which can be confusing and deceiving. 

Now, SSLF frame settings is now a child of the Bloom frame settings. In addition, there's a warning box directly in the override if any of the frame setting (bloom or SSLF) is set to false preventing the bloom pass to be executed without the users knowledge. 

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/6e522a47-79f2-49fe-aeaf-d9d4166467ef)

In addition, since SSLF writes directly on the bloom buffer and uses the uberpass bloom to be rendered, if bloom intensity is set to zero, it won't render. So in the SSLF override there's now a message warning you that one or more bloom override prevents SSLF to be rendered. 

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/16320f54-ad23-4338-a5c7-9e4f0788db65)

Lastly, I took the opportunity with this PR to add the frame settings check wherever it made sense in the current override to avoid further users confusions (cf [commit](https://github.cds.internal.unity3d.com/unity/unity/pull/51061/commits/32f1369d6710394879f23d2f24f84b73a68710cb)).