[VFX] Incorrect mesh rendered when using multimesh output and exposed submesh mask
When implementing instancing support for exposed meshes and textures, we missed one specific case.
If we use a multimesh output with a exposed submesh mask that comes from a expression tree that can be optimized, the expression will not be registered properly in runtime mode.
if the VFX is compiled in edition mode, the issue will not be visible.