[VFX] Fix Custom HLSL collection with diverging branches
Issue introduced at https://github.cds.internal.unity3d.com/unity/unity/pull/48953 (landed in 6000.0.6f1)

To illustrate the problem, here a case which doesn't cause any problem:
<img width="1088" alt="image" src="https://media.github.cds.internal.unity3d.com/user/42/files/364fc89f-9e77-4795-b82e-bc25f5718a72">

Failing case:
<img width="1085" alt="image" src="https://media.github.cds.internal.unity3d.com/user/42/files/7cfdad6d-ff4b-42b3-aafe-a4fb6c62854d">

Initially, I changed the workflow to collect IHLSLCodeHolder & GraphicsBufferUsage per context but the resolution was incomplete. If two end expressions were sharing the same ancestor, only the first evaluation is able to collected needed data. It's fine when we are catching this globally but this is wrong when the information must be computed per context.

After several back and forth, I finally decided to extract the expression collection from compilation and introduce a preprocess.

In the end, even if it forces to browse the expression graph twice, it separates concern and prevent side issues. Additionally, it removes the former `collectedData` in all compile functions.

🎁 I added `CollectPerContextData` flag to reserve this graph browsing to a single case.