[VFX] Fix Volumetric Fog Output with multiple outputs
When using multiple Volumetric Fog Outputs in a single system, the buffers `outputBuffer` and `maxSliceCount` were shared between the outputs, which was incorrect. It was leading to GPU hangs on some platforms, and to an excessive amount of slices being drawn on all platforms.

This PR makes sure that each output is working on its own set of buffers.

Renderdoc screenshots of the draw calls:
 - Before : the quad count is the same of both draw calls and is the sum of the two required quad count for each output ->Wrong
![image](https://media.github.cds.internal.unity3d.com/user/2768/files/58ce7d7c-b618-4428-8c8e-31d5e4ac2f78)
 - After : the quad count is different for each draw call and matches the number of required slices for each output -> Correct
![image](https://media.github.cds.internal.unity3d.com/user/2768/files/532e4d8d-3a4f-43ac-b4cb-2fcab81ca758)