Fix URP pass name handling
This is a continuation of https://github.cds.internal.unity3d.com/unity/unity/pull/50405

The goal is to have identical URP render pass names in all our profiling tools (RG viewer, Frame Debugger and Profiler).
The names are now also identical between RG and non-RG/Compatibility passes.

This PR fixes a number of potential bugs with regards to setting ScriptableRenderPass.profilerSampler to null, it fixes incorrect profiler scopes (inside the static method shared between RG and non-RG), improves more pass names using the naming convention from the previous PR and could have a very minor performance benefit in release build.   

In this PR, we separately store the passName and the profileSampler because these samplers can be null in the release build. Therefore, we can't just use sampler.name. By using the ScriptableRenderPass properties (passName, profileSampler) these names are kept in sync automatically.

The default name for a (user created) sub-class of a ScriptableRenderPass is now more descriptive, using the sub-class name instead of the base class.

The default RenderFeature that users create from the menu is updated to using this. 

The profile samplers being null in a release build was introduced by this PR as an optimization https://github.cds.internal.unity3d.com/unity/unity/pull/41824

However, in URP, many samplers are created differently, leading to limited performance gain, and inconsistency with some samplers being null and some not in the release build. This PR improves this so more samplers are null in a release build. Potentially gaining around 0.1ms main thread CPU in release on low end CPUs. 

Before and after this PR (no change), URP, ProfileSamplers references are kept in a few different places. Depending on that, the samplers can be nullified in the release build. 
- ScriptableRenderPass.profileSampler: used by RG and non-RG, in this PR, when using RG, they are nullified
- ProfileSampler.Get(URPProfileId): same pattern as with HDRP, these are nullified by the change from Julien
- RenderGraph.m_DefaultProfilingSamplers: created automatically when adding a pass without a sampler, these are nullified
- Other places of the URP code (outside of a pass): these are NOT nullified.

Likely, part of the optimization for HDRP came from the dictionary .TryGetValue with large pass dictionaries. The samplers kept outside of dictionary don't have this cost. So it likely will not improve performance to clean up these. 

Since some samplers can be null in the release build, it's bad practice (errors!) to use sampler.name.
However, the earlier PR to nullify this might have introduced issues already. To avoid issues with user project that do this (very low chance), in this PR, ScriptableRenderPass.profileSampler is only nullified in RenderGraph mode (not Compatibility).

This makes it a bit more complicated to ensure that the pass name and the sampler name (shown in profiler) don't get out of sync. However, it's confusing for users to see different names in different tools and complicates matching them during analysis. Hence the changes in this PR. 

The RenderGraph add pass methods (eg AddRasterPass, ...) have overloads for both passing a passName and a sampler, and one with only a passName. The use of passing only the passName seems recommended for users since they can't go out of sync, RG creates a new profiler with that name, and will nullify it in a release build, and users don't have access to the profilerSampler. In non release builds, there might be a small perf overhead of this practice since the dictionary can become large.